backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic
passlib[bcrypt]
python-multipart
requests
openai